Split oversized issue 3586 helpers
